core: firewall: differentiate configuration API
Add a firewall_conf structure to differentiate individual firewall
queries from firewall configuration that groups inseparable firewall
queries. A firewall configuration is expected to be set as a whole
and could group queries to different firewall controllers.

As a firewall configuration is composed of one or more queries: the
"[conf_name]-access-conf" property relies on the same cells as
"access-controllers".

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>