core: mm: strip the tag of VA when check_pa_matches_va
When CFG_TEE_CORE_DEBUG=y and CFG_MEMTAG=y, core will crash, if there are
some modules call virt_to_phys() or core_vbuf_is() and so on.
Because the MM can not find map by the tagged VA, so strip the tag of
VA when check_pa_matches_va().

Suggested-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: wentao.sun <wentao.sun@amlogic.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>