mk/compile.mk: remove duplicate dtc flags
When a .dts is compiled into a .dtb, the flags: '-I dts -O dtb
-Wno-unit_address_vs_reg' are present several times on the command line.
It is visible when building with "make ... V=1".

Fix this by not appending to DTC_FLAGS but instead use DTC_FLAGS as is
from the environment and add the specific flags once to the compile
command.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>