mk/compile.mk: separate rule to produce preprocessed dts files
When compiling a .dts into a .dtb, there is a pre-preprocessing phase
where the C preprocessor is run on the .dts to produce a .pre.dts, then
the device tree compiler is called to generate the .dtb. The two phases
are currently bundled into a single makefile recipe, which is not ideal.

Extract the preprocessing commands into a separate recipe to address
this concern. This allows to do for example:

 make PLATFORM=stm32mp1-157A_DK1 \
      out/arm-plat-stm32mp1/core/arch/arm/dts/stm32mp157a-dk1.pre.dts

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>