ta: remoteproc: add command to release remote processor firmware
This patch introduces a new command `TA_RPROC_CMD_RELEASE_FW` to release
the firmware images and associated resources of a remote processor.

This command can be called by the non-secure context to release the loaded
firmware if the remoteproc is not running.

The command can be used, for instance, in case an error occurs in the
non-secure context between the load request and the start of the
remote processor. Calling the command allows releasing the loaded firmware
and going back to the REMOTEPROC_OFF state.

This command should also be called after the stop of the
remote processor to release associated resources.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>