drivers: stm32_hsem: fix HSEM group CID filtering configuration
Move the check on HSEM group CID filtering configuration consistency
before the continuing if the CID filtering is disabled for the first
semaphore of the group.

The HSEM group CID filtering was also broken as it didn't allow more
than one CID to access a semaphore group. Fix this by using the semaphore
bit-field in the RIFPROT macro to determine what are the processors that
should be authorized to take the semaphores.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Fixes: ec9aa1a4114f ("drivers: add RIF support driver for HSEM")