libutils: util.h: assert size is a power of 2 in ROUNDUP_DIV()
Assert value of the rounding argument of ROUNDUP_DIV() is a power
of 2 as expected by the macro implementation.

Add ROUNDUP2_DIV() as an alias for ROUNDUP_DIV, for consistency with
the existing ROUNDxxx2() macros.

By the way, fix the indentation in the macro implementation for
consistency of the header file implementation.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>