core: imx: Enable ELE by default on supported platforms
The ELE provides the HUK on the SoCs which have it, so having support
for it on these SoCs is really prudent.

The ELE has been disabled by default in [1] and [2] because of:

| There is only one MU to communicate with ELE, which cannot be
| dedicated on OP-TEE side all the time. There may be ELE services
| running on Linux side, which can cause conflict with OP-TEE.

However, the ELE is currently only used for implementation of
plat_rng_init() and tee_otp_get_hw_unique_key(). The former is used only
during initialisation and the latter might be used during runtime, but
the ELE is actually only accessed once and the HUK is then cached for
later calls.

Enable the ELE by default again to make the config for i.MX8ulp and i.MX9
more useful.

[1] commit: 29b4cb6ef06f ("core: imx: disable ELE support on imx8ulp, imx93 by default")
[2] commit: 136cc65fc184 ("core: imx: disable ELE support on i.MX91 by default")

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>