plat-rockchip: increase FDT max size to 384KiB on all Aarch64 supported SoCs
Increase the maximum size of the FDT to 384KiB in sync with Trusted
Firmware-A since TF-A commit [1] available since v2.13 (May 2025). This
limit is applicable to all Rockchip SoCs supported by TF-A.

Prior to that commit in TF-A, we had 0x20000 (double the default of the
current OP-TEE OS default) since commit [2] available since v2.4 (Nov
2020).

This allows us to pass and parse the FDT within OP-TEE as the default
64KiB really isn't enough nowadays (especially if one takes into account
FDT with symbols enabled for FDTO support), otherwise OP-TEE OS panics
at:
E/TC:0   init_external_dt:827 Invalid Device Tree at 0x8a2690: error -3

Technically, rk322x should also be able to support that but I don't have
access to a device based on that SoC.

This fixes OP-TEE OS panics on PX30 and RK3399.

Link: https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/ab99dce4b7c8473d5bcb8c833bd410ab87b1e801%5E%21/ [1]
Link: https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/8109f738ffa79a63735cba29da26e7c2859977b5%5E%21/ [2]

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>