mk/compile.mk: add -undef and -D__DTS__ in dtb-cppflags
Commit 61b2d6e460f7 ("mk/compile.mk: add -Ulinux -Uunix to
dtb-cppflags") addresses some macro enabled by default by the
compiler that could impact the pre-processing of the DTS files.

Instead of un-defining each of such macros, use the flag '-undef'
and get rid of all of them in one shot.

Add the macro '__DTS__' specific for DTS pre-processing, as it is
done in Linux and later ported to U-Boot with [1].
This could help sharing include files between DTS and C code.

Link: https://github.com/u-boot/u-boot/commit/f53932addd31 [1]
Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>